Original commit
cleaner.js contains the javascript that is used to clean (sanitize)
HTML-formatted content for the purpose of making it safe for output in
javascript-enabled browser. This is the first version of the function
and inevitably contains omissions and gaps. Please contribute to make
it better. testrun.html contains a simple test application that can be
used to test the functionality.